backend/requirements.txt
# ============================================
# MEMORIA × ECHO — Production Dependencies
# ============================================

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.1
aiosqlite==0.20.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# AI / ML
openai==1.59.7
google-generativeai==0.8.4
transformers==4.47.1
torch==2.5.1
sentencepiece==0.2.0
langdetect==1.0.9
deep-translator==1.11.4
textblob==0.18.0

# Speech & NLP
SpeechRecognition==3.12.0
pyttsx3==2.98
nltk==3.9.1
spacy==3.8.4

# Real-time
websockets==14.1
python-socketio==5.12.0

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
Pillow==11.1.0
numpy==2.2.1

# Monitoring & Logging
loguru==0.7.3

# CORS
starlette==0.41.3
